+++ server/src/main/java/com/ones/admin/system/ApiResourceService.java
@@ -273,7 +273,7 @@ public class ApiResourceService {
             SystemMenuMapper menuMapper,
             SystemApiManifestSnapshotMapper manifestSnapshotMapper,
             ObjectMapper objectMapper,
-            @Value("${ones.version:v0.0.34}") String applicationVersion
+            @Value("${ones.version:v0.0.35}") String applicationVersion
     ) {
         this.requestMappingHandlerMapping = requestMappingHandlerMapping;
         this.permissionMapper = permissionMapper;